[LTB-32] Add tests and benchmarks (sketches of)
