+++ code/network/bench/Bench.hs
@@ -1,0 +1,3 @@
+-- | TBD.
+
+module Bench () where
